Consolidate 2_2 implementations
